+++ src/modes/windowed_mode.rs
@@ -344,7 +344,7 @@ pub fn winit_main(mut prog: Program) -> Result<(), &'static str> {
                         continue;
                     };
 
-                    eprintln!("Detected rate to be {}hz", milli_hz as f64 / 1000.0);
+                    eprintln!("Detected rate to be {}hz", milli_hz as f32 / 1000.0);
                     eprintln!("Note: Coffeevis relies on Winit to detect refresh rates.");
                     eprintln!("Refresh rate changed.\n");
                     *cmd = Command::FPSFrac(milli_hz);
